Implement useAdvertisingUrl.jsx
